add ModelRequest typing to getMatchingSubscription
