+++ client/src/app/domain/interfaces/model-request.ts
@@ -1,0 +1,30 @@
+import { Id } from '../definitions/key-types';
+
+export type FieldDescriptor = RelationFieldDescriptor | GenericRelationFieldDecriptor | StructuredFieldDecriptor;
+
+export interface Fields {
+    [field: string]: FieldDescriptor | null;
+}
+
+export interface HasFields {
+    fields: Fields;
+}
+
+export interface ModelRequest extends HasFields {
+    ids: Id[];
+    collection: string;
+}
+
+export interface GenericRelationFieldDecriptor extends HasFields {
+    type: 'generic-relation-list' | 'generic-relation';
+}
+
+export interface RelationFieldDescriptor extends HasFields {
+    type: 'relation-list' | 'relation';
+    collection: string;
+}
+
+export interface StructuredFieldDecriptor {
+    type: 'template';
+    values?: RelationFieldDescriptor | GenericRelationFieldDecriptor;
+}
